+++ rest_framework_json_api/utils.py
@@ -211,15 +211,15 @@ def get_related_resource_type(relation):
         relation_model = relation.model
     elif hasattr(relation, "get_queryset") and relation.get_queryset() is not None:
         relation_model = relation.get_queryset().model
-    elif (
-        getattr(relation, "many", False)
-        and hasattr(relation.child, "Meta")
-        and hasattr(relation.child.Meta, "model")
-    ):
-        # For ManyToMany relationships, get the model from the child
-        # serializer of the list serializer
-        relation_model = relation.child.Meta.model
-    else:
+    elif hasattr(relation, "child_relation"):
+        # For ManyRelatedField relationships, get the model from the child relationship
+        try:
+            return get_related_resource_type(relation.child_relation)
+        except AttributeError:
+            # Some read only relationships fail to get it directly, fall through to
+            # get via the parent
+            pass
+    if not relation_model:
         parent_serializer = relation.parent
         parent_model = None
         if isinstance(parent_serializer, PolymorphicModelSerializer):